Fixed circular dependency between dataHandler and localExec
