+++ tools/python/boutiques/exporter.py
@@ -3,7 +3,7 @@
 import json
 import os
 import uuid
-from boutiques.localExec import loadJson
+from boutiques.util.utils import loadJson
 from boutiques.logger import raise_error
